requirements.txt
airtable-python-wrapper==0.15.3
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
oauthlib==3.2.2
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyairtable==3.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.10.0
pyparsing==3.2.0
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
